multikey tests
These tests are for exercising compatibility with the Go client. In
particular, we want to interact with users who have new-style sigchains,
with sibkeys and multiple PGP keys. "t_frank" is a user who started with
a PGP key generated on the site but then logged in with the Go client to
generate a second PGP key. "t_george" is a user who signed up on the Go
client and just used that to generate two PGP keys. We check that
encrypt works against all the recipient's keys, and that the implicit
`id` passes against fingerprint assertions.