fix foo://bar user loading
We used to cache some data after we resolved something like
twitter://malgorithms, so that a subsequent user load would be faster.
But now that we do more during user loading (like fetching Merkle tree
data), what we were caching was inadequate, and it was preventing server
loads that needed to happen. Fix this by just getting rid of the caching
step in @resolve_user_name().

https://github.com/keybase/keybase-issues/issues/1680