rewrite the _verify checks
The immediate issue we needed to fix was that we were trying to populate
@gpg_keys in the case where the user had just reset and therefore had a
null @merkle_data.eldest_kid. Then I realized that this would probably
behave badly in other ways immediately after a reset, because has_links
is true. Instead of worrying about how many links the user has,
explicitly allow KeyOwnershipError in the case of the user's own
sigchain. (It should be impossible to trigger this after signing a
link in the latest subchain.)