update kbpgp and libkeybase
Pulls in a fix for https://github.com/keybase/kbpgp/issues/83.

We are now merging all versions of the same PGP key and ignoring revoked
subkeys. This isn't a very safe thing to be doing in general (someone
who's compromised your revoked subkey could try to make signatures with
it) but it solves the problem of users revoking a subkey today and
retroactively breaking their old sigchain links. Two things protect us
here:

1) We're in the middle of implementing exact key pinning, so that we
won't need to do any key merging for signatures made after that time.
This will be the Right Thing To Do.
2) The server still validates all new links against the most recent
exact key the user uploaded, so in the meantime an attacker exploiting
this would need the server's cooperation.

Also pulls in a fix for https://github.com/keybase/kbpgp/issues/80,
fixing a class of PGP keys we were failing to import.