stop asserting invitation code length
Now that we generate bip-0039 codes, their lengths are variable. Also
we've issues temporary promotional invitation codes before (like
"shityeahtwitter" or whatever it was) which probably never worked in the
node client. Rather than checking against a range of lengths that we
might violate in the future, just drop the check here.